removed repetitive code
